// src/assets/lib/family-chart.d.ts
declare interface Datum {
  id: string;
  rels: {
    father: string;
    mother: string;
    spouses: string[];
    children: string[];
  };
  data: {
    firstName: string;
    lastName: string;
    birthday: string;
    avatar: string;
    gender: string;
  },
  main?: boolean
}